// src/app/committee/lib/types.ts
export interface Member {
  id: string;
  firstName?: string;
  lastName?: string;
  name?: string;
  nickname?: string;
  department?: string;
  position?: string;
  room?: string;
  email?: string;
  instagram?: string;
  bio?: string;
  photoURL?: string;
  photos?: string[];
  order?: number;
}